+++ src/metrics/npa/ruby.rs
@@ -17,7 +17,7 @@ impl Npa for RubyCode {
     ) {
         use Ruby::*;
 
-        if Self::is_func_space(node) && stats.is_disabled() {
+        if opens_container_space::<Self>(node) && stats.is_disabled() {
             stats.is_class_space = true;
         }